Separate sso handlers from auth
